modem: chat: streamline codebase
Adjust struct alignment and use fast integer for better performance.

+++ subsys/modem/modem_chat.c
@@ -40,7 +40,7 @@ static void modem_chat_log_received_command(struct modem_chat *chat)
 	uint16_t log_buffer_pos = 0;
 	uint16_t argv_len;
 
-	for (uint16_t i = 0; i < chat->argc; i++) {
+	for (uint_fast16_t i = 0; i < chat->argc; i++) {
 		argv_len = (uint16_t)strlen(chat->argv[i]);
 
 		/* Validate argument fits in log buffer including termination */
@@ -420,7 +420,7 @@ static void modem_chat_parse_save_match(struct modem_chat *chat)
 static bool modem_chat_match_matches_received(struct modem_chat *chat,
 					      const struct modem_chat_match *match)
 {
-	for (uint16_t i = 0; i < match->match_size; i++) {
+	for (uint_fast8_t i = 0; i < match->match_size; i++) {
 		if ((match->match[i] == chat->receive_buf[i]) ||
 		    (match->wildcards == true && match->match[i] == '?')) {
 			continue;
@@ -435,9 +435,9 @@ static bool modem_chat_match_matches_received(struct modem_chat *chat,
 static bool modem_chat_parse_find_match(struct modem_chat *chat)
 {
 	/* Find in all matches types */
-	for (uint16_t i = 0; i < ARRAY_SIZE(chat->matches); i++) {
+	for (uint_fast16_t i = 0; i < ARRAY_SIZE(chat->matches); i++) {
 		/* Find in all matches of matches type */
-		for (uint16_t u = 0; u < chat->matches_size[i]; u++) {
+		for (uint_fast16_t u = 0; u < chat->matches_size[i]; u++) {
 			/* Validate match size matches received data length */
 			if (chat->matches[i][u].match_size != chat->receive_buf_len) {
 				continue;
@@ -461,7 +461,7 @@ static bool modem_chat_parse_find_match(struct modem_chat *chat)
 
 static bool modem_chat_parse_is_separator(struct modem_chat *chat)
 {
-	for (uint16_t i = 0; i < chat->parse_match->separators_size; i++) {
+	for (uint_fast8_t i = 0; i < chat->parse_match->separators_size; i++) {
 		if ((chat->parse_match->separators[i]) ==
 		    (chat->receive_buf[chat->receive_buf_len - 1])) {
 			return true;
@@ -473,7 +473,7 @@ static bool modem_chat_parse_is_separator(struct modem_chat *chat)
 
 static bool modem_chat_parse_end_del_start(struct modem_chat *chat)
 {
-	for (uint8_t i = 0; i < chat->delimiter_size; i++) {
+	for (uint_fast16_t i = 0; i < chat->delimiter_size; i++) {
 		if (chat->receive_buf[chat->receive_buf_len - 1] == chat->delimiter[i]) {
 			return true;
 		}
@@ -534,9 +534,9 @@ static void modem_chat_on_command_received_resp(struct modem_chat *chat)
 static bool modem_chat_parse_find_catch_all_match(struct modem_chat *chat)
 {
 	/* Find in all matches types */
-	for (uint16_t i = 0; i < ARRAY_SIZE(chat->matches); i++) {
+	for (uint_fast16_t i = 0; i < ARRAY_SIZE(chat->matches); i++) {
 		/* Find in all matches of matches type */
-		for (uint16_t u = 0; u < chat->matches_size[i]; u++) {
+		for (uint_fast16_t u = 0; u < chat->matches_size[i]; u++) {
 			/* Validate match config is matching previous bytes */
 			if (chat->matches[i][u].match_size == 0) {
 				chat->parse_match = &chat->matches[i][u];
@@ -692,7 +692,7 @@ static void modem_chat_process_byte(struct modem_chat *chat, uint8_t byte)
 
 static bool modem_chat_discard_byte(struct modem_chat *chat, uint8_t byte)
 {
-	for (uint8_t i = 0; i < chat->filter_size; i++) {
+	for (uint_fast16_t i = 0; i < chat->filter_size; i++) {
 		if (byte == chat->filter[i]) {
 			return true;
 		}
@@ -704,7 +704,7 @@ static bool modem_chat_discard_byte(struct modem_chat *chat, uint8_t byte)
 /* Process chunk of received bytes */
 static void modem_chat_process_bytes(struct modem_chat *chat)
 {
-	for (uint16_t i = 0; i < chat->work_buf_len; i++) {
+	for (uint_fast16_t i = 0; i < chat->work_buf_len; i++) {
 		if (modem_chat_discard_byte(chat, chat->work_buf[i])) {
 			continue;
 		}
@@ -868,7 +868,7 @@ int modem_chat_run_script_async(struct modem_chat *chat, const struct modem_chat
 	}
 
 	/* Validate script commands */
-	for (uint16_t i = 0; i < script->script_chats_size; i++) {
+	for (uint_fast16_t i = 0; i < script->script_chats_size; i++) {
 		if ((script->script_chats[i].request_size == 0) &&
 		    (script->script_chats[i].response_matches_size == 0) &&
 		    (script->script_chats[i].timeout == 0)) {
